my_app/requirements.txt
asgiref==3.7.2
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
Django==3.2.20
django-allauth==0.54.0
django-background-tasks==1.2.5
django-compat==1.0.15
django-crontab==0.7.1
django-environ==0.10.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.7.0
oauthlib==3.2.2
packaging==23.1
psycopg2==2.9.7
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
waitress==2.1.2
wincertstore==0.2
zipp==3.15.0
